+++ objectbox_test/test/sync_test.dart
@@ -29,6 +29,11 @@ void main() {
   late TestEnv env2;
   int serverPort = 9999;
 
+  /// Returns a test server URL with the current [serverPort].
+  serverUrl() {
+    return 'ws://127.0.0.1:$serverPort';
+  }
+
   setUp(() {
     env = TestEnv('sync');
     store = env.store;
@@ -47,7 +52,7 @@ void main() {
   // lambda to easily create clients in the tests below
   SyncClient createAuthenticatedClient(
           Store s, List<SyncCredentials> credentials) =>
-      Sync.clientMultiCredentials(s, 'ws://127.0.0.1:$serverPort', credentials);
+      Sync.clientMultiCredentials(s, serverUrl(), credentials);
 
   SyncClient createClient(Store s) =>
       createAuthenticatedClient(s, [SyncCredentials.none()]);
@@ -253,6 +258,29 @@ void main() {
         SyncCredentials.jwtRefreshToken('refresh-token'),
         SyncCredentials.jwtCustomToken('custom-token')
       ]);
+    });
+
+    test('SyncClient filter variables', () {
+      final filterVariables = {
+        'test-var-1': 'test value 1',
+        'test-var-2': 'test value 2'
+      };
+      SyncClient client = Sync.client(
+          store, serverUrl(), SyncCredentials.none(),
+          filterVariables: filterVariables);
+      addTearDown(() {
+        client.close();
+      });
+
+      client.putFilterVariable('test-var-2', 'test value 2');
+      client.removeFilterVariable('test-var-2');
+      client.putFilterVariable('test-var-2', '');
+      client.removeAllFilterVariables();
+
+      expect(
+          () => client.putFilterVariable('', 'value'),
+          throwsA(isA<ArgumentError>().having((e) => e.message, 'message',
+              contains('Filter variables must have a name'))));
     });
 
     group('Server tests using sync-server in PATH', () {